Added pragma case_sensitive ON to take advantage of indexes.
The current searches of the dictionary are not using the index at all.

Queries that use LIKE cannot reliably use an index. The queries of the
form LIKE "text%" can use it, but only when case sensitive is on. This
is not an issue for Japanese.

See https://stackoverflow.com/questions/8584499/sqlite-should-like-searchstr-use-an-index